requirements.txt
flask==3.1.0
google-genai==1.14.0
pillow==11.1.0
gunicorn==23.0.0
python-dotenv==1.1.0
